rag-backend/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
asyncpg
python-dotenv
pydantic[email]
python-multipart
pyyaml
